+++ CHANGELOG.md
@@ -7,6 +7,7 @@ InsertDate Changelog
 - Fixed a bug where the user config was appended to the popup every time. (#13)
 - Updated default prompt configurations
 - Added several command palette commands to access configuration files
+- Removed microseconds from "iso" and "unix" format
 
 
 v0.5.4 (2014-02-19)